savadkoohi

+++ FinalProject/src/finalproject/Bullet.java
@@ -18,8 +18,8 @@ import javax.imageio.ImageIO;
  */
 public class Bullet {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     public BufferedImage[] bullets;
     public int boxNumber;
     private int numberOfBox;
@@ -35,19 +35,19 @@ public class Bullet {
 
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
     public int boxCrashedNumber() {
